feat: signal for sending the created user account activation message
This signal was created and connected to the use case of the 'users' module that allows the registration of users with the 'searcheruser' role, now every time a registration is completed successfully, a sent message will be sent to the user's email. along with instructions on how to activate your account.